+++ Assets/VRM10/Editor/Components/SpringBone/VRM10SpringBoneJointEditor.cs
@@ -59,12 +59,12 @@ namespace UniVRM10
             m_showJointSettings = EditorGUILayout.Foldout(m_showJointSettings, "Joint Settings");
             if (m_showJointSettings)
             {
-                LimitBreakSlider(m_stiffnessForceProp, 0.0f, 4.0f, 0.0f, Mathf.Infinity);
-                LimitBreakSlider(m_gravityPowerProp, 0.0f, 2.0f, 0.0f, Mathf.Infinity);
+                Vrm10EditorUtility.LimitBreakSlider(m_stiffnessForceProp, 0.0f, 4.0f, 0.0f, Mathf.Infinity);
+                Vrm10EditorUtility.LimitBreakSlider(m_gravityPowerProp, 0.0f, 2.0f, 0.0f, Mathf.Infinity);
                 EditorGUILayout.PropertyField(m_gravityDirProp);
                 EditorGUILayout.PropertyField(m_dragForceProp);
                 EditorGUILayout.Space();
-                LimitBreakSlider(m_jointRadiusProp, 0.0f, 0.5f, 0.0f, Mathf.Infinity);
+                Vrm10EditorUtility.LimitBreakSlider(m_jointRadiusProp, 0.0f, 0.5f, 0.0f, Mathf.Infinity);
             }
 
             if (serializedObject.ApplyModifiedProperties())
@@ -141,53 +141,6 @@ namespace UniVRM10
             }
 
             return jointIndex == (spring.Joints.Count - 1);
-        }
-
-        /// <summary>
-        /// スライダーと数値入力で限界値の違う、所謂「限界突破スライダー」を作成する
-        /// `EditorGUILayout.PropertyField` の代替として利用する
-        /// </summary>
-        private static void LimitBreakSlider(SerializedProperty property, float sliderLeft, float sliderRight, float numberLeft, float numberRight)
-        {
-            var label = new GUIContent(property.displayName);
-            var currentValue = property.floatValue;
-
-            var rect = EditorGUILayout.GetControlRect();
-
-            EditorGUI.BeginProperty(rect, label, property);
-
-            rect = EditorGUI.PrefixLabel(rect, label);
-
-            // slider
-            {
-                EditorGUI.BeginChangeCheck();
-
-                var sliderRect = rect;
-                sliderRect.width -= 55.0f;
-                rect.xMin += rect.width - 50.0f;
-
-                var clampedvalue = Mathf.Clamp(currentValue, sliderLeft, sliderRight);
-                var sliderValue = GUI.HorizontalSlider(sliderRect, clampedvalue, sliderLeft, sliderRight);
-
-                if (EditorGUI.EndChangeCheck())
-                {
-                    property.floatValue = sliderValue;
-                }
-            }
-
-            // number
-            {
-                EditorGUI.BeginChangeCheck();
-
-                var numberValue = Mathf.Clamp(EditorGUI.FloatField(rect, currentValue), numberLeft, numberRight);
-
-                if (EditorGUI.EndChangeCheck())
-                {
-                    property.floatValue = numberValue;
-                }
-            }
-
-            EditorGUI.EndProperty();
         }
 
         void OnSceneGUI()
