+++ Assets/VRM10/Editor/Vrm10EditorUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b75cffea6a5f03418a9a698a32c15cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
